# swagger/cadastroClientes.yaml
swagger: '2.0'
info:
  title: Cadastro de Clientes
  description: Operações de cadastro e busca de clientes
  version: 1.0.0
host: localhost:8080
schemes: 
   - http
definitions:
...
  Endereco:
    type: object
    properties:
      id:
        type: number
        format: long
      rua:
        type: string
      cep:
        type: string
      bairro:
        type: string
      cidade:
        type: string
      estado:
        type: string
      complemento:
        type: string
      referencia:
        type: string
      cliente:
        type: object
...
paths:
...
  /api/enderecos:
    post:
      summary: Cadastra um endereço na base de dados
      parameters:
        - in: body
          name: dto
          required: true
          schema:
            $ref: '#/definitions/Endereco'
      produces:
        - application/json
      responses:
        '200':
          description: o cep foi encontrado e o endereço foi salvo
        '400':
          description: o cadastro não foi finalizado
  /api/enderecos/{id}:
    put:
      summary: Atualiza um endereço cadastrado na base de dados
      parameters:
        - in: path
          name: id
          type: number
          format: long
          description: Id do endereço a ser atualizado
          required: true
        - in: body
          name: dto
          required: true
          schema:
            $ref: '#/definitions/Endereco'
      produces:
        - application/json
      responses:
        '201':
          description: Endereço foi atualizado com sucesso
        '400':
          description: Houve uma falha na atualização de endereço
  /api/enderecos/filtrar:
    get:
      summary: Busca endereços pelo id do cliente
      parameters:
        - in: query
          name: idCliente
          type: number
          format: long
          required: true
          description: id do cliente que deseja buscar endereco
      produces:
        - application/json
      responses:
        '201':
          description: Busca foi concluída
        '400':
          description: Id do cliente informado na busca não foi encontrado